Move io package to feature/stream
