+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/feature/stream/StreamService.java
@@ -17,7 +17,7 @@
  * (C) 2021 tesshucom
  */
 
-package com.tesshu.jpsonic.service;
+package com.tesshu.jpsonic.feature.stream;
 
 import static org.springframework.web.bind.ServletRequestUtils.getBooleanParameter;
 import static org.springframework.web.bind.ServletRequestUtils.getIntParameter;
@@ -36,12 +36,18 @@ import com.tesshu.jpsonic.domain.system.PreferredFormatSheme;
 import com.tesshu.jpsonic.feature.auth.jwt.JWTAuthenticationToken;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
-import com.tesshu.jpsonic.io.PlayQueueInputStream;
 import com.tesshu.jpsonic.persistence.api.entity.MediaFile;
 import com.tesshu.jpsonic.persistence.api.entity.PlayQueue;
 import com.tesshu.jpsonic.persistence.api.entity.Player;
 import com.tesshu.jpsonic.persistence.core.entity.User;
+import com.tesshu.jpsonic.service.AudioScrobblerService;
+import com.tesshu.jpsonic.service.MediaFileService;
+import com.tesshu.jpsonic.service.PlaylistService;
+import com.tesshu.jpsonic.service.SearchService;
+import com.tesshu.jpsonic.service.SecurityService;
+import com.tesshu.jpsonic.service.StatusService;
 import com.tesshu.jpsonic.service.StatusService.TransferStatus;
+import com.tesshu.jpsonic.service.TranscodingService;
 import com.tesshu.jpsonic.service.TranscodingService.VideoTranscodingSettings;
 import com.tesshu.jpsonic.service.scanner.WritableMediaFileService;
 import com.tesshu.jpsonic.util.PlayerUtils;
