+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/feature/stream/DownloadControllerTest.java
@@ -17,7 +17,7 @@
  * (C) 2021 tesshucom
  */
 
-package com.tesshu.jpsonic.controller;
+package com.tesshu.jpsonic.feature.stream;
 
 import static com.tesshu.jpsonic.util.PlayerUtils.now;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
